Various functionality and robustness improvements
Updated the CREATE problem-type to allow data to be imported after the CREATE
TABLE operations have been run.  This allows subsequent tests to use provided
data, and to exercise the schema in some very basic ways.  This required
updating ProblemType to take the assignment name, so that files can be looked
up.

The FUNCTION problem-type can deal with converting between different kinds of
numeric values, when students decide to do unusual things.

The PROCEDURE problem-type has a bit more flexibility for examining database
state before and after a procedure is run.

The DB tools code attempts to clear the cursor's result-sets when students put
in multiple answers for queries.  This helps improve robustness.